requirements.txt
selenium==4.9.0
webdriver-manager-4.0.2
